refactor main.rb and app.rb and created person and rental modules
